+++ SilverSolutions1151/Middleware/Services/Interfaces/ISoldStockService.cs
@@ -4,7 +4,7 @@ namespace SilverSolutions1151.Middleware.Services.Interfaces
 {
     public interface ISoldStockService
     {
-        int GeSoldByDate(DateTime endDate);
+        decimal GeSoldByDate(DateTime endDate);
         List<Manufacture> GetSoldStockBalance(DateTime endDate);
     }
 }
